requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
pdfminer.six>=20221105
pycountry>=22.3.5
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
jupyter>=1.0.0
nbconvert>=7.0.0
